microservices/py-automation/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pandas==2.2.3
requests==2.32.3
sqlalchemy==2.0.35
mysql-connector-python==9.1.0
